requirements.txt
pandas
pandas-ta>=0.3.14b
python-dotenv
pybit
